// src/app/lib/auth/interfaces/auth.interface.ts
export interface SignInInterfaceI {
    email:              string;
    password:           string;
}

export interface SignUpInterfaceI {
    firstname:          string;
    lastname:           string;
    email:              string;
    password:           string;
    confirmpassword:    string;
}